chore: convert request_queue_collection too
